+++ src/extension.ts
@@ -6,6 +6,7 @@ import { AuthenticationWebviewViewProvider } from './activitybar/authenticationW
 import { SqlTreeDataProvider } from './activitybar/sqlTreeDataProvider';
 import { SETTING_ACTIVE_CONNECTION, SETTING_CONNECTIONS, SETTING_PINNED_OBJECTS } from './services/connections';
 import * as commands from './extensionCommands';
+import { initTokenHosts } from './services/tokenHosts';
 import { FsqlCompletionItemProvider } from './language/fsqlCompletionItemProvider';
 import { FsqlDocumentSemanticTokensProvider } from './language/fsqlDocumentSemanticTokensProvider';
 import { FsqlInlayHintsProvider } from './language/fsqlInlayHintsProvider';
@@ -45,6 +46,7 @@ export function getExtensionUri(): Uri {
 export function activate(context: ExtensionContext) {
 
 	extensionUri = context.extensionUri;
+	initTokenHosts(context.globalState);
 
 	let queryResultsWebviewMapping: Map<string, ResultsRender> = new Map<string, ResultsRender>();